+++ modules/frontend/config.go
@@ -90,7 +90,6 @@ func (cfg *Config) RegisterFlagsAndApplyDefaults(string, *flag.FlagSet) {
 			ConcurrentRequests:    defaultConcurrentRequests,
 			TargetBytesPerRequest: defaultTargetBytesPerRequest,
 			Interval:              5 * time.Minute,
-			RF1ReadPath:           false,
 			Exemplars:             false, // TODO: Remove?
 			MaxExemplars:          100,
 		},